add node-api for plugin sdk - easier node access on ui thread of the plugin.
